third_party/blink/web_tests/accessibility/name-calc-group-inside-treeitem.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
  li { list-style-type: none; }
</style>
<ul role="tree">
  <li role="treeitem" id="parent1">
    Parent1
    <ul role="group" id="group1">
      <li role="treeitem" id="treeitem1">
        Child1
      </li>
    </ul>
  </li>
</ul>

<div role="tree">
  <span role="treeitem" id="parent2">
    Parent2
    <span role="group">
      <span role="treeitem">
        Child2
      </span>
    </span> 
  </span>  
</div>


<script>
test(function(t) {
    var axParent1 = accessibilityController.accessibleElementById("parent1");
    assert_equals(axParent1.name.trim(), "Parent1");
}, "Accessible name does not include display:block descendants with nameFrom:author");

test(function(t) {
    var axParent2 = accessibilityController.accessibleElementById("parent2");
    assert_equals(axParent2.name.trim(), "Parent2");
}, "Accessible name does not include display:inline descendants with nameFrom:author");
</script>
